requirements.txt
# Sparser
loguru
requests
ujson

# Telegram
aiogram
gotify

# Vk
python-dotenv
vkbottle
